// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as auth from "../auth.js";
import type * as calls from "../calls.js";
import type * as channels from "../channels.js";
import type * as directMessages from "../directMessages.js";
import type * as http from "../http.js";
import type * as lib_authz from "../lib/authz.js";
import type * as messages from "../messages.js";
import type * as presence from "../presence.js";
import type * as serverMembers from "../serverMembers.js";
import type * as servers from "../servers.js";
import type * as signals from "../signals.js";
import type * as typingIndicators from "../typingIndicators.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  calls: typeof calls;
  channels: typeof channels;
  directMessages: typeof directMessages;
  http: typeof http;
  "lib/authz": typeof lib_authz;
  messages: typeof messages;
  presence: typeof presence;
  serverMembers: typeof serverMembers;
  servers: typeof servers;
  signals: typeof signals;
  typingIndicators: typeof typingIndicators;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};